// src/lib/client/models/IdeologySectionConditioner.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { IdeologyConditioner } from './IdeologyConditioner';
export type IdeologySectionConditioner = {
  readonly uuid: string;
  /**
   * Name of this condition rule.
   */
  name: string;
  /**
   * Description of the logic applied in this rule.
   */
  description?: string | null;
  /**
   * List of values that satisfy this condition (e.g. ['Spain']). Must match values in the conditioner.
   */
  condition_values?: any;
  readonly conditioner: IdeologyConditioner;
};